optimizations:
- true/false/null regexes:
With the current set of regexes, we don't need to check for a word break before; it won't change any results. Removing this \b check improves lexer speed slightly.

- regex loop
After matching the whitespace pattern, the next character won't be whitespace, so we can proceed to the next pattern.

This seems to give ~10% faster parsing overall.